backend/requirements.txt
Flask==2.2.5
Werkzeug==2.2.3
flask-sqlalchemy==3.0.2
flask-migrate==4.0.4
flask-login==0.6.2
flask-cors==3.0.10
flask-session==0.6.0
python-dotenv==0.19.0
requests==2.31.0
PyJWT==2.8.0
google-auth==2.28.1
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0 